+++ packages/manager/apps/billing/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.20](https://github.com/ovh/manager/compare/@ovh-ux/manager-billing-app@0.2.19...@ovh-ux/manager-billing-app@0.2.20) (2025-02-10)
+
+**Note:** Version bump only for package @ovh-ux/manager-billing-app
+
+
+
+
+
+## [0.2.19](https://github.com/ovh/manager/compare/@ovh-ux/manager-billing-app@0.2.18...@ovh-ux/manager-billing-app@0.2.19) (2025-02-06)
+
+**Note:** Version bump only for package @ovh-ux/manager-billing-app
+
+
+
+
+
 ## [0.2.18](https://github.com/ovh/manager/compare/@ovh-ux/manager-billing-app@0.2.17...@ovh-ux/manager-billing-app@0.2.18) (2025-02-03)
 
 **Note:** Version bump only for package @ovh-ux/manager-billing-app