+++ packages/manager/apps/iplb/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.10](https://github.com/ovh/manager/compare/@ovh-ux/manager-iplb-app@3.1.9...@ovh-ux/manager-iplb-app@3.1.10) (2025-02-06)
+
+**Note:** Version bump only for package @ovh-ux/manager-iplb-app
+
+
+
+
+
 ## [3.1.9](https://github.com/ovh/manager/compare/@ovh-ux/manager-iplb-app@3.1.8...@ovh-ux/manager-iplb-app@3.1.9) (2025-02-03)
 
 **Note:** Version bump only for package @ovh-ux/manager-iplb-app