+++ packages/manager/apps/nasha/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.22](https://github.com/ovh/manager/compare/@ovh-ux/manager-nasha-app@4.0.21...@ovh-ux/manager-nasha-app@4.0.22) (2025-02-06)
+
+**Note:** Version bump only for package @ovh-ux/manager-nasha-app
+
+
+
+
+
+## [4.0.21](https://github.com/ovh/manager/compare/@ovh-ux/manager-nasha-app@4.0.20...@ovh-ux/manager-nasha-app@4.0.21) (2025-02-06)
+
+**Note:** Version bump only for package @ovh-ux/manager-nasha-app
+
+
+
+
+
 ## [4.0.20](https://github.com/ovh/manager/compare/@ovh-ux/manager-nasha-app@4.0.19...@ovh-ux/manager-nasha-app@4.0.20) (2025-02-03)
 
 **Note:** Version bump only for package @ovh-ux/manager-nasha-app